requirements.txt
python-telegram-bot==21.6
ollama==0.3.3
apscheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.2
faster-whisper==1.0.3
google-api-python-client==2.136.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
